Changed card selection algorithm. Once a card is selected it is now removed from the card pool, eliminating the need for the do/while loop.
Changed Bane card selection algorithm. Now generates a list of possible candidates from the card pool, and selects the Bane card from the list. If the list is empty (because all 2 or 3 cost cards from the chosen sets have already been selected to be in the game), throws an exception.

Added a Restrictions class to the Dominion class. This class will be where any of restrictions a user wants to impose on their game (like no attacks) will go. Currently only the restriction available is to abide by Alchemy's recommendation that only 3-5 Alchemy cards be in the game (provided at least one other set is chosen).

Added method to implement Alchemy recommendation restriction.